Integral error eliminated (atol & rtol 10-5) in model map generation stage (prior culprit)
Added 1..15 degree square cubature schemes for more granular control of integral

Switched Primary test harness to googletest from doctest

+++ xtsrcmaps/fits/fits.hxx
@@ -31,7 +31,7 @@ struct TablePars {
 auto read_irf_pars(std::string const&, std::string) -> std::optional<TablePars>;
 
 auto write_src_model(std::string const&              filename,
-                     Tensor<float, 4> &         model_map,
+                     Tensor<double, 4> &         model_map,
                      std::vector<std::string> const& srcs) -> void;
