+++ xtsrcmaps/model_map/model_map.scale_map_by_solid_angle.cxx
@@ -36,9 +36,9 @@ dir_diff(std::array<T, 3> const& L, std::array<T, 3> const& R) -> T {
 }
 
 auto
-solid_angle(Fermi::Tensor<float, 3> const& points,
-            Fermi::SkyGeom<float> const&   skygeom) -> Fermi::Tensor<float, 2> {
-    Fermi::Tensor<float, 2> phi(points.extent(0), points.extent(1));
+solid_angle(Fermi::Tensor<double, 3> const& points,
+            Fermi::SkyGeom<double> const& skygeom) -> Fermi::Tensor<double, 2> {
+    Fermi::Tensor<double, 2> phi(points.extent(0), points.extent(1));
     for (size_t h = 0; h < points.extent(0); ++h) {
         for (size_t w = 0; w < points.extent(1); ++w) {
             // Adapted from FermiTools CountsMap.cxx:612 and FitsImage.cxx:108
@@ -51,18 +51,18 @@ solid_angle(Fermi::Tensor<float, 3> const& points,
             auto const D
                 = skygeom.pix2dir({ points[h, w, 0] + 1.f, points[h, w, 1] });
 
-            float dOmega1
+            double dOmega1
                 = dir_diff(A, B) * dir_diff(A, D)
-                  * norm(cross(normalize(std::array<float, 3> {
+                  * norm(cross(normalize(std::array<double, 3> {
                                    A[0] - B[0], A[1] - B[1], A[2] - B[2] }),
-                               normalize(std::array<float, 3> {
+                               normalize(std::array<double, 3> {
                                    A[0] - D[0], A[1] - D[1], A[2] - D[2] })));
 
-            float dOmega2
+            double dOmega2
                 = dir_diff(C, B) * dir_diff(C, D)
-                  * norm(cross(normalize(std::array<float, 3> {
+                  * norm(cross(normalize(std::array<double, 3> {
                                    C[0] - B[0], C[1] - B[1], C[2] - B[2] }),
-                               normalize(std::array<float, 3> {
+                               normalize(std::array<double, 3> {
                                    C[0] - D[0], C[1] - D[1], C[2] - D[2] })));
 
             phi[h, w] = 0.5 * (dOmega1 + dOmega2);
@@ -72,15 +72,15 @@ solid_angle(Fermi::Tensor<float, 3> const& points,
 }
 
 void
-Fermi::ModelMap::scale_map_by_solid_angle(Tensor<float, 4>&     model_map,
-                                          SkyGeom<float> const& skygeom) {
+Fermi::ModelMap::scale_map_by_solid_angle(Tensor<double, 4>&     model_map,
+                                          SkyGeom<double> const& skygeom) {
     size_t const Ns = model_map.extent(0);
     size_t const Nh = model_map.extent(1);
     size_t const Nw = model_map.extent(2);
     size_t const Ne = model_map.extent(3);
 
     /* Tensor3d const init_points = get_init_centers(Nh, Nw); */
-    Tensor<float, 3> init_points(Nh, Nw, 2);
+    Tensor<double, 3> init_points(Nh, Nw, 2);
     for (size_t h = 0; h < Nh; ++h) {
         for (size_t w = 0; w < Nw; ++w) {
             init_points[h, w, 0] = 1. + h;
